+++ src/main/java/com/college/excursion/dao/impl/HaractDaoImpl.java
@@ -7,7 +7,7 @@ import com.college.excursion.dao.HaractDao;
 import com.college.excursion.model.Groupp;
 import com.college.excursion.model.Haract;
 @Repository
-public class HaractDaoImpl extends BaseDaoImpl<Haract,Long> implements HaractDao {
+public class HaractDaoImpl extends BaseDaoImpl<Haract,Integer> implements HaractDao {
 
 	public HaractDaoImpl() {
 		super(Haract.class);
